+++ RedPandaIDE/parser/parserutils.h
@@ -114,6 +114,7 @@ enum class StatementKind  {
     Parameter,
     Block,
     Lambda,
+    Label,
     UserCodeSnippet,  // user code template
     Keyword, // keywords
     KeywordType, //keywords for type (for color management)